@@ -102,16 +102,16 @@ public final class TestUtil {
             if (item == null) {
                 return null;
             }
-            if (item instanceof BigDecimal) {
-                return (BigDecimal) item;
-            } else if (item instanceof Long) {
-                return BigDecimal.valueOf((Long) item);
-            } else if (item instanceof Integer) {
-                return BigDecimal.valueOf((Integer) item);
-            } else if (item instanceof Double) {
-                return BigDecimal.valueOf((Double) item);
-            } else if (item instanceof Float) {
-                return BigDecimal.valueOf((Float) item);
+            if (item instanceof BigDecimal decimal) {
+                return decimal;
+            } else if (item instanceof Long long1) {
+                return BigDecimal.valueOf(long1);
+            } else if (item instanceof Integer integer) {
+                return BigDecimal.valueOf(integer);
+            } else if (item instanceof Double double1) {
+                return BigDecimal.valueOf(double1);
+            } else if (item instanceof Float float1) {
+                return BigDecimal.valueOf(float1);
             } else {
                 return BigDecimal.valueOf(item.doubleValue());
             }
